+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Add `paragraph` props on `ProjectCard` component.
+
 ## [20.10.0] - 2018-05-04
 
 Feature: